Converter: cast an untyped-constant shift to its resolved type (CS0034)
When the shifted (left) operand of a shift is an untyped constant — `1 << k` — Go gives the
whole shift the type it assumes from context (e.g. uintptr when compared with a uintptr), but
the bare C# literal makes the result `int`, which then can't compare or combine with the typed
operand: `uintptrVal < (1 << k)` → CS0034 ("operator '<' is ambiguous on nuint and int"). This
is the runtime/internal/math MulUintptr `a|b < 1<<(4*goarch.PtrSize)` pattern.

In convBinaryExpr's shift handling, when the left operand is an untyped numeric constant and the
shift's resolved type is a concrete non-`int` integer, cast the shift result to that type:
`u < (uintptr)(1 << (int)(k))`. Greens runtime/internal/math (with the namespace-collision and
MaxUintptr fixes).

Churn: re-baselined ShiftPrecedenceUnsigned + StdLibInternalAbi (both iteration-13 shift tests) —
the untyped-const shifts now carry their type (e.g. `(uint16)(1<<15)`, `(byte)(1<<0)`); values and
output unchanged, both still pass. ConversionStrategies updated.

+++ src/Tests/Behavioral/StdLibInternalAbi/type.cs
@@ -492,7 +492,7 @@ public static readonly ΔChanDir InvalidDir = 0;
 }
 
 [GoRecv] public static bool IsVariadic(this ref ΔFuncType t) {
-    return (uint16)(t.OutCount & ((1 << (int)(15)))) != 0;
+    return (uint16)(t.OutCount & ((uint16)(1 << (int)(15)))) != 0;
 }
 
 [GoType] partial struct PtrType {
@@ -529,15 +529,15 @@ public static ж<byte> Data(this ΔName n, nint off) {
 }
 
 public static bool IsExported(this ΔName n) {
-    return (byte)((n.Bytes.val) & ((1 << (int)(0)))) != 0;
+    return (byte)((n.Bytes.val) & ((byte)(1 << (int)(0)))) != 0;
 }
 
 public static bool HasTag(this ΔName n) {
-    return (byte)((n.Bytes.val) & ((1 << (int)(1)))) != 0;
+    return (byte)((n.Bytes.val) & ((byte)(1 << (int)(1)))) != 0;
 }
 
 public static bool IsEmbedded(this ΔName n) {
-    return (byte)((n.Bytes.val) & ((1 << (int)(3)))) != 0;
+    return (byte)((n.Bytes.val) & ((byte)(1 << (int)(3)))) != 0;
 }
 
 public static (nint, nint) ReadVarint(this ΔName n, nint off) {
@@ -602,14 +602,14 @@ public static ΔName NewName(@string n, @string tag, bool exported, bool embedde
     byte bits = default!;
     nint l = 1 + nameLenLen + len(n);
     if (exported) {
-        bits |= (byte)((1 << (int)(0)));
+        bits |= (byte)((byte)(1 << (int)(0)));
     }
     if (len(tag) > 0) {
         l += tagLenLen + len(tag);
-        bits |= (byte)((1 << (int)(1)));
+        bits |= (byte)((byte)(1 << (int)(1)));
     }
     if (embedded) {
-        bits |= (byte)((1 << (int)(3)));
+        bits |= (byte)((byte)(1 << (int)(3)));
     }
     var b = new slice<byte>(l);
     b[0] = bits;
